Change dir structure and fixed doc review comments
