+++ module_utils/network/ios/config/bgp/process.py
@@ -26,11 +26,11 @@
 # USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
 from ansible.module_utils.network.common.utils import to_list
-from ansible.module_utils.cisco_ios.config import ConfigBase
-from ansible.module_utils.cisco_ios.config.bgp import get_bgp_as
-from ansible.module_utils.cisco_ios.config.bgp.neighbor import BgpNeighbor
-from ansible.module_utils.cisco_ios.config.bgp.address_family import BgpAddressFamily
-from ansible.module_utils.cisco_ios.config.bgp.network import BgpNetwork
+from ansible.module_utils.network.ios.config import ConfigBase
+from ansible.module_utils.network.ios.config.bgp import get_bgp_as
+from ansible.module_utils.network.ios.config.bgp.neighbor import BgpNeighbor
+from ansible.module_utils.network.ios.config.bgp.address_family import BgpAddressFamily
+from ansible.module_utils.network.ios.config.bgp.network import BgpNetwork
 
 
 class BgpProcess(ConfigBase):
