+++ module_utils/network/ios/config/bgp/timer.py
@@ -25,7 +25,7 @@
 # LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 # USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
-from ansible.module_utils.cisco_ios.config import ConfigBase
+from ansible.module_utils.network.ios.config import ConfigBase
 
 
 class BgpTimer(ConfigBase):
